requirements.txt
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.0.10
python-dotenv>=0.19.0
pytube>=12.1.0
youtube-transcript-api>=0.4.1
openAI>=0.27.0